pyproject.toml
[tool.poetry]
name = "ocr-pdf-converter"
version = "0.1.0"
description = "OCR PDF Converter using Tesseract"
authors = ["Borhan <borhan@example.com>"]
readme = "README.md"
packages = [{include = "ocr_pdf_converter", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
pymupdf = "^1.24.0"
typer = "^0.12.3"
rich = "^14.0.0"
Pillow = "^12.0.0"
streamlit = "^1.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ocr-pdf = "ocr_pdf_converter.cli:main"
